OGM-1064 Move unsupported methods to AbstractGroupingByEntityDialect
